Bug fix: insertion sort requires working on a slice too, not Vec.
By using Vec, Rust kept ownership which was making the code FFI crash.
See http://stackoverflow.com/questions/32238751/rust-ffi-how-to-fix-this-segmentation-fault-when-passing-an-array-of-int

Because we have a slice and not a vec, we can't insert() and remove()
elements from the vector as this changes its allocation. Instead,
use swap() on the element to place it at the proper location. This might
not be as optimal of a solution (it might have been better to swap all
elements at once). But this works.